Fix destroy() in Velox aggregates
Aggregates are expected to free their variable length storage in stringAllocator_ of their RowContainer. This adds the required frees. In an accumulator contains a container, like F14* or std::vector or std::priority_qeue, the container's destructor must be called even if the container is empty.

We check that the HashStringAllocator for variable length content is
empty at destruction. We call destroy() of the accumulators innstead
of relying on a clear of the HashStringAllocator.

In the event of spilling, a fraction of the accumulators is
freed. The above check make sure that the right data gets freed in
spilling. Under other conditions the clear of the HashStringAllocator
would in principle work for cases where any containers exclusively
allocate from the HashStringAllocator and are otherwise trivially
destructible.

Fixes destruction of xpriority queues in TopNRowNumber.cpp. This is analogous to the case with aggregates.

The global check in RowContainer for enforcing destruct of aggregates
is not enabled.