Support longer fuzzer runs when changing sensitive files. (#6009)
Summary:
Creates dynamic config that runs a longer fuzzer job when changes are made to sensitive files. This was previously reverted because CCI couldnt run scheduled jobs with dynamic config. Now that scheduled jobs have been moved , we can re-enable this.

When we detect changes to sensitive files (say some file in velox/exec/...) then the presto fuzzer job is run for an hour during the PR.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6009

Reviewed By: xiaoxmeng

Differential Revision: D48087776

Pulled By: kgpai

fbshipit-source-id: f30114af4237e21d500570c2824b5b998f581a99