Remove capacity enforcement from MemoryManager (#5414)
Summary:
This PR removes capacity tracking and enforcement in MemoryManager, because of the previously added functionality of all allocator's ability to track and enforce capacity. Tests have been modified to accommodate the change.
This PR also fixed a bug in MallocAllocator/MmapAllocator on reservation consistency. Under certain conditions in calls like allocateNonContiguous and allocateContiguous, the freed collateral bytes might not be known by the caller MemoryPool, hence the pool might not release the freed bytes.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5414

Reviewed By: xiaoxmeng

Differential Revision: D47035604

Pulled By: tanjialiang

fbshipit-source-id: 5689da15590d5d08cd1fcffafb04d4a37e71cb28