Try next larger free list if head is no fit
HashStringAllocator can hit a worst case if items are freed in
dcreasing order of size inside a free list bucket. If one looks for
the top size in the bucket, one must travel to the end to find it and
it may still not be found. We move the head of the fre list to the
place where we stopped the search for the last time. In this way, a
situation with all small before all the large will not have to
repeatedly go through the small.

So we quit looking after a few tries and move to the larger bucket where any hit will be large enough.