Add GPU memory and API basics (#6015)
Summary:
Adds an arena for managing device or universal memory for query execution on accelerators. Code modified from MmapArena.*. The std::maps will be replaced by a SIMD tree for performance later.

Adds a wrapper for Cuda concepts. The Cuda headers will not be included in CPU code and the Velox headers will not be included in kernel code. Adds a test for the patttern of making a specialized stream for an interface between the C++ and Cuda code so that the Cuda headers are not included in C++.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6015

Reviewed By: Yuhta

Differential Revision: D48115432

Pulled By: oerling

fbshipit-source-id: 1081e812e3ec2cfa55d733c20cf87289dd4375fe