Add Cuda Event, Metrics and Utilities (#6096)
Summary:
Encapsulates Cuda event. Tests timing of callbacks and stream-to-stream synchronization with Cuda events. Adds a generic test for multi-stream operation from one host stream with options for prefetch, synchronization and callbacks.

A host callbakc from Cuda stream is about 100us. A two stream synchronization on device is about 3 us. A 10 stream synchronization on device is about 40 us. The cost to launch a kernel from a stream is about 3 us if it is enqueued beforehand.

Times with GForce 3080 ad 16x PCIE 4.0

Adds a header for common cub applications, initially with a flag array to indices of set positions conversion. More thread block utilities to folllow.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/6096

Reviewed By: Yuhta

Differential Revision: D48305411

Pulled By: oerling

fbshipit-source-id: 260bdfa064a2df40fbff0ed728b31c55bda85265