Fix race condition in SharedArbitratorTest.reclaimFromJoinBuilder* tests reading/writing to QueryConfig
Summary:
These tests use AssertQueryBuilder to update the QueryConfig in the QueryCtx object, this runs in one thread.  Unfortunately the same QueryCtx may be used to run another query in a second thread.  This can lead to the QueryConfig being read and written at the same time leading to undefined behavior.

The fix is to just update the QueryConfig before we start running any of the threads.

Differential Revision: D48538101

fbshipit-source-id: bfad28da54eedc578836cb446a6e072fa6be96cc