Add mechanism to run non-streaming (blocking) wave operator (#6221)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6221

Divide wave operators in driver into multiple pipelines; each pipeline starts with either a source operator (e.g. `Values`, `TableScan`) or blocking operator (e.g. `Aggregation`, `HashBuild`).  The rest of the pipeline should be streaming (no mutable state is kept in the operator and current output is solely depending on the current input).

At end of each pipeline, we generate some result vector.  If it is the last pipeline, we convert the result back to CPU vector.  Otherwise we create a `WaveVector` and enqueue it into the starting operator (guaranteed to be blocking) of next pipeline.  Once all the streams finishes enqueuing the results, we flush the enqueued vectors to build up internal states of the blocking operator.

The output of blocking operator is fetched in the same way as a normal source operator by polling `canAdvance`.

Differential Revision: D48606213

fbshipit-source-id: 32b1d50df3bd3e820d84c99270f5b09c30086950