Fix: avoid suppression of VeloxRuntimeError (#6145)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6145

VeloxRuntimeError should always be propagated to the top level and not suppressed. At least not
in expression eval or function evaluation.
This diff fixes several places where the error was supressed such as casting, repeat, remote ..etc .
Its not trivial to write a unit test for each, but lets make sure we  review the code carefully.

Reviewed By: kagamiori

Differential Revision: D48438243

fbshipit-source-id: 583d4a3c968fa257ed71dccd017c856223619c4a