Fix: Make sure switch expr generates vectors that can be dereferenced at rows where errors are thrown (#6243)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6243

When switch expression is evaluated and the output is complex type, for indices where
an error happens, nothing is written on the output vector.

This can result in a vector that is not serializable and cant be dereferenced for those
indices which can result in seg-fault when someone tries to access the data.

We have a rule that vectors should be always serializable for their size, furthermore,
this could lead to error if any expression later on is reading the result for the indices
were we did throw, this diff fixes the issue here. see the issue discussion for more details
This fix:https://github.com/facebookincubator/velox/issues/5878

following fix will update the cast as well to avoid being performed on rows that have errors.

Reviewed By: bikramSingh91

Differential Revision: D48480584

fbshipit-source-id: fa34362d14473f65fdc4f6c97947f5824609d5ce