Fix PeeledEncoding to peel off inner-most Lazy layer
Summary:
For vectors like dictionary(lazy(flat)), PeeledEncoding doesn't peel off the inner-most lazy layer. This is because the peeling loop doesn't update peeledVector[i] with the loaded lazy vector in the peeling loop, so after the peeling process stops, the remaining code returns peeledVector[i] that is a lazy vector. This caused a crash in CastExpr that expect simple vector after peeling but got a lazy vector.

This diff fixes this issue by replacing peeledVector[i] with the loaded vector once the loading happens.

Differential Revision: D48814919

fbshipit-source-id: 6bb2ea921e38031bdb796ae7f69b24200131110c