Support accessing sub-fields by index (#6253)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6253

This diff adds a new ITypedExpr DereferenceTypedExpr which accesses a subfield of a Row by field index
(unlike FieldAccessTypedExpr which only allows accessing fields by name).

This is particularly useful for cases where the RowType was defined without names and so they all default to
empty string, making it impossible to access any field but the first one, but is generally useful for any cases
where field names are for some reason not unique.

I hope that we can eventually use FieldAccessTypedExpr specifically for accessing top level columns, while
DereferenceTypedExpr is used for accessing subfields.  This will simplify logic in the compiler, for example,
where today we have to check if a FieldAccessTypedExpr is an input column or not.  In general, I split the two
classes for this reason as well as making it easy to tell what's available (Dereference always has index, while
FieldAccess always has name).

Reviewed By: mbasmanova

Differential Revision: D48665326

fbshipit-source-id: b38e3e55102756f08760d4aacd0b97be308b4f71