Fix: CheckDuplicateKeys should not receive rows with null keys. (#6341)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/6341

This fix https://github.com/facebookincubator/velox/issues/6301.

CheckDuplicateKeys calls toString on the base vector value to generate an error message.
If the key is null this can cause a Velox runtime error in the case of accessing a dictionary
vector with added nulls and 0 sizes base vector!

- This function does not need to process null keys, calls should remove rows with null keys
before calling this function.
- Users should not access wrappedIndex(index) when the top level dict is null at index.
- The diff also change MapFromEntries and Map functions to remove rows with nulls keys before
calling CheckDuplicateKeys. TransformKeys will be updated in a separate PR.

Differential Revision: D48797877

fbshipit-source-id: fd218fa4a415b9383a85a9d170f0bf1750cb809b