Attribute lazy load times to table scan
When recording addInput, getOutput and finish times for Operators,
subtracts the lazy load time accrued to the operator since last timing
update. Adds the lazy load CPU and wall time to the source operator of
the pipeline. Tracks lazy load with both CPU and wall time since this
must be compatible with the CPU and wall time tracking in operator
timing.

The runtime stats for lazy load are still kept at the level of the
operator triggering the load. These times do not contribute to the
operator's time as reported in addInputTiming and related.