Move native Parquet reader to fbthrift
Moving the thrift dependency from Parquet reader from thrift to
fbthrift. Fbthrift is already used for Prestissimo and remote function
execution, and the symbols conflict with thrift when linked to the same
binary. FBthrift is also supposed to provide better performance and a
more flexible API based on IOBuf, which allow us to remove some of the
memory copies we do currently.