Add different Cuda memory types and copy meter (#6508)
Summary:
Adds allocators for device and host memory. Adds a test for different types of host to device transfer. Memory is transferred by prefetch, copy or staging into pinned and explicyt copy into device memory. A reduction is made on the device and the result is transferred back to host. This is done with overlapping transfers from one or mire threads with different numbers of simulated columns and different batch sizes.


Reviewed By: Yuhta

Differential Revision: D49137804

Pulled By: oerling