Change date.h to better support int32_t days from epoch
From date_diff function, we find two bugs of the date.h lib
in supporting the DATE type with days from epoch in the range
of int32_t. Fixes are as follows

1. class year: change that type of 'year' member variable to
int (from short) to avoid overflow.

2. year_month_day::from_days function: use type long for
an intermediate variable to avoid overflow.

date.h comes from https://github.com/HowardHinnant/date/issues

This is a quick fix to unblock. Ideally, we want to open an issue and ask
the maintainers of the 'date' library to fix.