Fix the linking problem of simdjson
Summary:
When velox was used as a third-party library and `SIMDJsonExtractor`
was used, it failed when running json function tests. We found that
`-DSIMDJSON_THREADS_ENABLED=1` was not configured when generating
libvelox_functions_json.a. We fix it by changing "simdjson" to
"simdjson::simdjson" in target_link_libraries.

Fixes https://github.com/facebookincubator/velox/issues/6564