Track number of rows processed in Window::getOutput
Streaming window needs the number of rows processed in
the getOutput() call for the result vector. Not all received
rows might be assigned to a partition yet. Infact the last
partition is always open until noMoreInput is received.
So getOutput() can only process all closed partitions
which is less than the number of rows input.