Ensure RowVector::resize also resizes the children. (#6543)
Summary:
Currently when we call resize on a rowVector we do not resize its children. This creates problems for example in ContainerRowSerde where we resize the parent row (but the children still have the old size) and the new rows are not set to null. In general ensuring children are of same size as parent in resize helps to ensure vector is in a valid state.


Reviewed By: laithsakka, DanielMunozT

Differential Revision: D49241838

Pulled By: kgpai