Rename dataType parameter to fileType for the ColumnReaders
As of now, the file data types were named as "fileType_" in
SelectiveColumnReader, but "nodeType_" in ColumnReader. Moreover, the
the corresponding parameter in the constructors were named as "dataType"
or "nodeType", which are obscure and not self-explanatory. This commit
renames these entities to "fileType" uniformly, so that the class member
names and parameter names match and the code is eaiser to read.