Fix crash in ExprCompiler (#6634)
Summary:
When expression contains array_sort function call, rewriteExpression rewrites it
and returns a different expression. compileRewrittenExpression used to store a
pointer to rewritten expression in F14FastMap, which went out of scope and
resulting in a dangling pointer stored in F14FastMap. Whenever, F14Map needed
to check that entry it causes a crash.

A fix is to store an std::shared_ptr to rewritten expression in the
Scope struct.

New test reproduces the crash.


Reviewed By: xiaoxmeng

Differential Revision: D49419393

Pulled By: mbasmanova