Fix ArbitraryBuffer::empty()
`ArbitraryBuffer::empty()` returns true if there is no buffered page,
including two cases: no page exists, and only an end-marker exists.
For scenarios where the arbitrary buffer is marked as "no-more-data"
but there are still buffered pages in it, it should return false.

Because it is not a public API, this fix will not affect current users.