Avoid copying result in FieldReference when propagating nulls (#6632)
Summary:

In `FieldReference::evalSpecialForm`, we copy the child vector unconditionally (because it is referenced in both parent vector and `result` vector) in order to add nulls on it .  This is particularly inefficient for nested structs.  Fix this by allowing adding null in-place if the child is uniquely owned by the parent.

There is another place in `Expr::evalWithNulls` that we also make a copy of the column, but it is done only once for each top level column, so it is relatively less expensive comparing to `FieldReference` which is also copying on each internal vector.  Still there is some chance to further optimize it.

Reviewed By: oerling

Differential Revision: D49400857