Fast Path for Scalar Flat Map
Flat map decoding for scalar values was extremely bad. The main reason
is generating too many intermediate structures (list of copy sources
and targets) and calling virtual functions (BaseVector::copy*) to
copy.

Now, do all scalar payloads as follows: Generate a row-wise bitmap
marking all the items to copy from per-key vectors to direct map
values or keys. Generate this column-wise: Every selected row where
the key is present sets a bit. The init sets the bits with a stride of
distinct keys. So if 100 distinct keys and all rows are selected and
all rows have a value for the key, set bits 0, 100, 200 etc to mark
that the value is needed. Counting the bits produces the size of the
key/valyue vector for the direct map.

As a second pass, loop through the bits in row-wise order. Every set
bit means that a value needs to be copied from the row/column of the
bit to the next place in the direct map's keys/values.

If the keys or values are strings, there can be constant or dictionary encoding in the values. These are flattened out in the copy.

For scalar key/value types an extra optimization could be to do a
gather load - vector store for keys and values. We'll see if we do
this later. For now, the acceleration is 4-5x.

Adds a test for string-string flat map to exercise string keys and
values inline and not inline. Adds a utility to make not in line
strings.