Alternate API to Register FileSystem
File system is initialised when the fist invocation to filesystem is made. As
not all classes have access to the connector config, they call getFileSystem
with nullptr in pace of connector config. The side-effect of this being, if
this happens to be first invocation of filesystem, the filesystem instance
gets initialised incorrectly.

Some filesystems require do not require config object to be passed on each
invocation of getFileSystem, this PR adds a new registerFileSystem API.
Once config is passed to registerFileSystem, the config object will be re-used
for all the subsequent calls to getFileSystem.