Rename rows_ to ranges_ in detail::Destination
The `rows_` member of `detail::Destination` is misleading because it is
actually a list of `IndexRange` (an `IndexRange` represents a row-range).
In an `IndexRange`, there may be only one row, but it may also include
multiple consecutive rows(in the scenario of only one partition).

To improve readability and avoid misunderstandings, `rows_` is renamed
to `ranges_` and `row_` is renamed to `rangeIdx_`.

No functional change in the PR.