Ensure WarmStorageFileSystem created with FileSystemTimeoutConfig.
Summary:
We use warm_storage::getDefaultFileSystemTimeoutConfig(), which
means 'no timeout' for anything.
At the same time we create custom FileSystemTimeoutConfig, which timeouts
conming from the config file (with default values). But it wasn't used.
So, use it now and log the values.

Also, emsure HdfsFileSystem get config in the same way as other file systems.

Differential Revision: D49606585