Support UNKNOWN type in RowContainer's compare functions (#6733)
Summary:

As the title says, this change adds support for the UNKNOWN type in RowContainer's compare functions.  This is trivial as the comparison already supports NULLs.

I also explicitly marked Opaque types as unsupported.

I considered adding a fast path in the untemplated compare functions for UNKNOWN (we can just return 0 directly), but the small performance benefit didn't seem worth the risk of introducing bugs (e.g. potentially missing a check in the templated compare functions) in the future.

Differential Revision: D49606066