Trace WS calls.
Summary:
Trace WS calls, so we can collect data on how many of them
are in-flight, max ms, etc.
The planned use of the traces:
1. In presto_cpp collect counters in the periodic task.
2. In presto_cpp create a server operation that will return current traces.

Why do we need this?
We've been observing WS deadlocks and hangs and need ways to debug this.

Note, that we aren't tracing other FileSystem implementations - we don't
need that now.

Differential Revision: D49674829