Spill all to disk to simplify query oom on output processing (#6764)
Summary:
In Meta internal arbitration and spilling test, we found if a query has
spilled, then it is very likely the remaining state in-memory takes non-trivial
amount of memory which is close to the query memory limit during the
output processing, and it is very likely causing oom during that stage and
we don't support spilling in the middle of output processing.
The followup is to (1) add spilling support during the output processing,
(2) add spilling support during the output processing with mixed of spilled
and in-memory state. And then remove this limitation.


Reviewed By: oerling

Differential Revision: D49698546

Pulled By: xiaoxmeng