Fix Timestamp::toTimezone for negative timestamps.
The value of `seconds_` in `Timepstamp` is the starting point of one
second, so when converting a `time_point` from milliseconds to seconds,
it should always be rounded down.

However, `std::chrono::duration_cast` will always round to 0, so for
negative numbers (times before 1970-01-01 00:00:00), the result will
be wrong(will differ from the correct result 1 second).