Fix: resize loaded vector to match lazy vector size. (#6663)
Summary:

LazyVector::ensureLoadedRows() used to be able to replace the input vector with a new vector.
This used to be done to ensure that the output vector is valid when we load less rows than the size
of the lazy vector.

This is problematic because the input to ensureLoadedRows() is not checked to be unique (hence other
references are not not updated) also if there any raw references they are not also updated.

It's hard to make sure there are no raw references obtained to the input of ensureLoadedRows.
For example, this code i invalid because children of row vector are not updated.

```
auto children =(decoded.base()->asUnchecked<RowVector>()) ->children();
DecodedVector decodedChild;
SelectivityVector childRows;
for (auto& child : children) {
decodedChild.decode(*child, baseRows, false);
ensureLoadedRowsImpl(child, decodedChild, baseRows, childRows);
}
```

while it looks easy to fix up there by using a reference instead of copying children pointers, there is
guarantee that those children or any nested vector where ensureLoadedRowsImpl will be called is
unique or does not have a row pointer.

To solve the issue, ensureLoadedRows is changed not to change the input vector, instead if we push the requires vector size to the loader. the loader would make sure the size of the loadedVector has the same size of the containing lazy that is passed to it.
This would keep the containing vectors valid without a need to change anything else.

Reviewed By: Yuhta, bikramSingh91

Differential Revision: D49461830