Fix null input handling in Reduce function
Reduce currently used the BaseVector's addNulls to add nulls at the end
to the localResult. This API is very basic and does not handle the variety
of cases that Expr::addNulls is designed to handle. For eg, if the target
vector is smaller than the input rows, or has a different encoding or is
empty. This change moves the Expr::addNulls() API to EvalCtx to make
it easily accessible to vector functions like Reduce without pulling in the
complete Expr header file and finally employing it to add nulls.

 Test Plan:
 Added unit tests