Fix flaky test for zombie task
When check `ASSERT_EQ(task.use_count(), 1)` is enforced, it is possible
that `waitForTaskFailure` has not finished, holding the task reference
and causing the test failure. Instead of checking task.use_count(), we
can rely on `QueryAssertions#waitForAllTasksToBeDeleted` to check if
the task is hanging at the end of the test.