De-flake tests that verify task failures (#6907)
Summary:

Tests that run multi-threaded tasks that are expected to fail sometimes crash.

```
F1005 01:30:34.221653 2215122 DefaultKeepAliveExecutor.h:145] Check failed: keepAliveCount > 0
```

This happens because the Task and the Executor are destroyed while some drivers are still running. A fix is to wait for all drivers to complete before allowing the Task to go out of scope.

Reviewed By: xiaoxmeng

Differential Revision: D49947841