compareMaterializedRows should compare unique row counts
Summary:
compareMaterializedRows in QueryAssertions currently checks that the number of rows in Velox and DuckDb
match, and that every row output by one appears in the other's results.

This is not strict enough, as query results may have duplicate rows.  compareMaterializedRows will miss the
case where the total number of rows is the same but the number of each distinct row is different.

The fix is to check that the counts match, not just that the row is present.

Differential Revision: D49895573